size of characters

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
-/** 
+/**
  * main - a prgram that prints size
- * Return: 0 (Success) 
+ * Return: 0 (Success)
  */
 int main(void)
 {
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float e;
 
-printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
+printf("size of a char: %lu byte(s)\n", sizeof(a));
+printf("size of an int: %lu byte(s)\n", sizeof(b));
+printf("size of a long int: %lu byte(s)\n", sizeof(c));
+printf("size of a long long int: %lu byte(s)\n", sizeof(d));
+printf("size of a float: %lu byte(s)\n", sizeof(e));
 return (0);
 }
